EditorWindow class changes
Improved the EditorWindow class structure and setter/getter methods.
Removed the public properties of the EditorWindow class and made them
private. Added access methods for the private properties and avoid
setters if not needed. Introduced the TextAction type in the Types
namespace to improve the access to the actions of the QPlainTextEdit
base class of my EditorArea class.